--- WIPProject/AzureCloudService/DatabaseConnectionServiceContent/diagnostics.wadcfgx
@@ -1,39 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<DiagnosticsConfiguration  xmlns="http://schemas.microsoft.com/ServiceHosting/2010/10/DiagnosticsConfiguration">
-  <PublicConfig>
-    <WadCfg>
-      <DiagnosticMonitorConfiguration overallQuotaInMB="4096">
-        <DiagnosticInfrastructureLogs scheduledTransferLogLevelFilter="Error"/>
-        <Logs scheduledTransferPeriod="PT1M" scheduledTransferLogLevelFilter="Error" />
-        <Directories scheduledTransferPeriod="PT1M">
-          <IISLogs containerName ="wad-iis-logfiles" />
-          <FailedRequestLogs containerName ="wad-failedrequestlogs" />
-        </Directories>
-        <WindowsEventLog scheduledTransferPeriod="PT1M" >
-          <DataSource name="Application!*[System[(Level=1 or Level=2 or Level=3)]]" />
-          <DataSource name="Windows Azure!*[System[(Level=1 or Level=2 or Level=3 or Level=4)]]" />
-        </WindowsEventLog>
-        <CrashDumps containerName="wad-crashdumps" dumpType="Mini">
-          <CrashDumpConfiguration processName="WaIISHost.exe"/>
-          <CrashDumpConfiguration processName="WaWorkerHost.exe"/>
-          <CrashDumpConfiguration processName="w3wp.exe"/>
-        </CrashDumps>
-        <PerformanceCounters scheduledTransferPeriod="PT1M">
-          <PerformanceCounterConfiguration counterSpecifier="\Memory\Available MBytes" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\Web Service(_Total)\ISAPI Extension Requests/sec" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\Web Service(_Total)\Bytes Total/Sec" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\ASP.NET Applications(__Total__)\Requests/Sec" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\ASP.NET Applications(__Total__)\Errors Total/Sec" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\ASP.NET\Requests Queued" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\ASP.NET\Requests Rejected" sampleRate="PT3M" />
-          <PerformanceCounterConfiguration counterSpecifier="\Processor(_Total)\% Processor Time" sampleRate="PT3M" />
-        </PerformanceCounters>
-      </DiagnosticMonitorConfiguration>
-    </WadCfg>
-    <StorageAccount></StorageAccount>
-  </PublicConfig>
-  <PrivateConfig>
-    <StorageAccount name="" key="" endpoint="" />
-  </PrivateConfig>
-  <IsEnabled>true</IsEnabled>
-</DiagnosticsConfiguration>
