--- WIPProject/DatabaseConnService/DatabaseConnService.svc
@@ -1,1 +1,0 @@
-﻿<%@ ServiceHost Language="C#" Debug="true" Service="DatabaseConnectionService.DatabaseConnService" CodeBehind="DatabaseConnService.svc.cs" %>
