--- WIPProject/DatabaseConnService/IDatabaseConn.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.Serialization;
-using System.ServiceModel;
-using System.ServiceModel.Web;
-using System.Text;
-using System.Data.SqlClient;
-
-namespace DatabaseConnectionService {
-
-    [ServiceContract]
-    public interface IDatabaseConn {
-        [OperationContract]
-        [WebInvoke(Method = "POST", ResponseFormat = WebMessageFormat.Json)]
-        string GetConnenctionToDB();
-
-        [OperationContract]
-        [WebInvoke(Method = "POST", ResponseFormat = WebMessageFormat.Json)]
-        string CheckLogin(string userName, string password);
-
-        [OperationContract]
-        [WebInvoke(Method = "POST", ResponseFormat = WebMessageFormat.Json)]
-        string AddNewUser(string userName, string password);
-    }
-}
