--- WIPProject/DatabaseConnService/Web.config
@@ -1,61 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-  <system.diagnostics>
-    <trace>
-      <listeners>
-        <add type="Microsoft.WindowsAzure.Diagnostics.DiagnosticMonitorTraceListener, Microsoft.WindowsAzure.Diagnostics, Version=2.8.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" name="AzureDiagnostics">
-          <filter type=""/>
-        </add>
-      </listeners>
-    </trace>
-  </system.diagnostics>
-  <system.web>
-    <customErrors mode="Off"/>
-    <compilation debug="true" targetFramework="4.6.2"/>
-  </system.web>
-  <system.serviceModel>
-    <services>
-      <service name="DatabaseConnectionService.DatabaseConnService">
-        <endpoint address="/DBConn" binding="basicHttpBinding" contract="DatabaseConnectionService.IDatabaseConn">
-        </endpoint>
-        
-        <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange">
-        </endpoint>
-      </service>   
-    </services>
-    <behaviors>
-      <serviceBehaviors>
-        <behavior>
-          <!-- To avoid disclosing metadata information, set the value below to false before deployment -->
-          <serviceMetadata httpGetEnabled="true"/>
-          <!-- To receive exception details in faults for debugging purposes, set the value below to true.  Set to false before deployment to avoid disclosing exception information -->
-          <serviceDebug includeExceptionDetailInFaults="true"/>
-        </behavior>
-      </serviceBehaviors>
-    </behaviors>
-    <serviceHostingEnvironment multipleSiteBindingsEnabled="true"/>
-  </system.serviceModel>
-  <system.webServer>
-    <modules runAllManagedModulesForAllRequests="true"/>
-    <!--
-        To browse web app root directory during debugging, set the value below to true.
-        Set to false before deployment to avoid disclosing web app folder information.
-    -->
-    <directoryBrowse enabled="true"/>
-  </system.webServer>
-
-	<runtime>
-		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-			<dependentAssembly>
-				<assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30AD4FE6B2A6AEED" culture="neutral"/>
-				<bindingRedirect oldVersion="0.0.0.0-8.0.0.0" newVersion="8.0.0.0"/>
-			</dependentAssembly>
-		</assemblyBinding>
-	</runtime>
-</configuration>
-<!--
-I changed Include Exeption details to false
-Endpoint = /DBConn
-Contract = IDatabaseConn
-Aervice Name = DatabaseConnService
--->
