+++ WIPProject/WCFServiceWebRole1/DatabaseConnService.svc
@@ -1,0 +1,1 @@
+﻿<%@ ServiceHost Language="C#" Debug="true" Service="DatabaseConnectionService.DatabaseConnService" CodeBehind="DatabaseConnService.svc.cs" %>
