+++ WIPProject/WCFServiceWebRole1/WebRole.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.WindowsAzure;
+using Microsoft.WindowsAzure.Diagnostics;
+using Microsoft.WindowsAzure.ServiceRuntime;
+
+namespace WCFServiceWebRole1 {
+    public class WebRole : RoleEntryPoint {
+        public override bool OnStart() {
+            // For information on handling configuration changes
+            // see the MSDN topic at https://go.microsoft.com/fwlink/?LinkId=166357.
+
+            return base.OnStart();
+        }
+    }
+}
