+++ WIPProject/WCFServiceWebRole1/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net462" />
+  <package id="Microsoft.Data.Edm" version="5.7.0" targetFramework="net462" />
+  <package id="Microsoft.Data.OData" version="5.7.0" targetFramework="net462" />
+  <package id="Microsoft.Data.Services.Client" version="5.7.0" targetFramework="net462" />
+  <package id="Newtonsoft.Json" version="8.0.2" targetFramework="net462" />
+  <package id="System.Spatial" version="5.7.0" targetFramework="net462" />
+  <package id="WindowsAzure.Storage" version="8.0.1" targetFramework="net462" />
+</packages>
